requirements.txt
geopandas
dash
folium
python-levenshtein
gunicorn
rq
schedule